refactor: introduce shared risk base
